backend/requirements.txt
pydantic==2.11.9
pydantic-ai>=0.0.15
langfuse==3.3.5
fastapi>=0.110.0
python-dotenv>=1.0
uvicorn[standard]>=0.24.0
faiss-cpu>=1.7.4
numpy>=1.26
sentence-transformers>=2.2.0
